function moved to timeutils
